# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Website John Services Motel API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/createService": {
            "post": {
                "tags": [
                    "Service"
                ],
                "summary": "Create",
                "description": "Creation",
                "operationId": "createService",
                "requestBody": {
                    "description": "Enregistrer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title_en",
                                    "title_fr",
                                    "description_fr",
                                    "description_en"
                                ],
                                "properties": {
                                    "title_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "rdc"
                                    },
                                    "title_fr": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "rdc"
                                    },
                                    "description_fr": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner kambale"
                                    },
                                    "description_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "rdc"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "rdc"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success"
                    },
                    "401": {
                        "description": "ysteme"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/updateService/{id}": {
            "post": {
                "tags": [
                    "Service"
                ],
                "summary": "Update",
                "description": "Modification",
                "operationId": "updateService",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Modifier",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title_en",
                                    "title_fr",
                                    "description_fr",
                                    "description_en"
                                ],
                                "properties": {
                                    "description_fr": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner services"
                                    },
                                    "description_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner services"
                                    },
                                    "title_en": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "winner kambale"
                                    },
                                    "title_fr": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "rdc"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "rdc"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "404": {
                        "description": "Service not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/deleteService/{id}": {
            "delete": {
                "tags": [
                    "Service"
                ],
                "summary": "Delete",
                "description": "Suppression",
                "operationId": "deleteService",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Slide deleted successfully"
                    },
                    "404": {
                        "description": "Slide not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/getSingleService/{id}": {
            "get": {
                "tags": [
                    "Service"
                ],
                "summary": "Afficher",
                "description": "Afficher ",
                "operationId": "getSingleService",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'approv",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}